+++ kube_dns/CHANGELOG.md
@@ -4,7 +4,7 @@
 ==================
 ### Changes
 
-* [IMPROVEMENT] Add metric `kubedns.request_duration.seconds.count`. See [#1341][]
+* [IMPROVEMENT] Add metrics `kubedns.request_count.count`, `kubedns.error_count.count` and `cachemiss_count.count`, alternative metrics submitted as monotonic\_counts. See [#1341][]
 
 1.2.0 / 2018-01-10
 ==================
@@ -16,7 +16,7 @@
 ==================
 ### Changes
 
-* [UPDATE] Update auto_conf template to support agent 6 and 5.20+. See [#860][]
+* [UPDATE] Update auto\_conf template to support agent 6 and 5.20+. See [#860][]
 
 1.0.0 / 2017-07-18
 ==================